src/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-dotenv==1.0.1
cassandra-driver==3.29.2
PyYAML==6.0.2
